requirements.txt
absl-py==0.11.0
adjustText==0.7.3
aiosignal==1.2.0
altgraph==0.17.4
amqp==2.6.1
anyjson==0.3.3
apipkg==1.5
appdirs==1.4.3
asgiref==3.3.1
astor==0.8.1
astroid==2.11.2
astunparse==1.6.3
attrs==20.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.3.0
bioinfokit==1.0.4
cached-property==1.5.2
cachetools==4.1.1
category-encoders==2.4.1
celery==4.4.2
certifi==2020.11.8
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.12
Circos==1.3.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cligj==0.7.2
cloudpickle==1.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
cycler==0.10.0
dask==2021.3.0
dataclasses==0.8
decorator==4.4.2
dill==0.3.1.1
distlib==0.3.6
Django==3.1.3
django-celery-results==2.0.0
django-extensions==3.1.0
django-filter==2.4.0
django-picklefield==3.0.1
django-rename-app==0.1.2
djangorestframework==3.12.2
drf-yasg==1.20.0
entrypoints==0.4
et-xmlfile==1.1.0
execnet==1.7.1
filelock==3.4.1
Fiona==1.8.22
flake8==3.8.4
flatbuffers==1.12
flower==0.9.5
frozenlist==1.2.0
future==0.18.3
gast==0.3.3
geopandas==0.9.0
globus-sdk==1.9.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
graphviz==0.19.1
grpcio==1.32.0
grpcio-tools==1.28.1
h5py==2.10.0
httpie==2.3.0
humanize==3.1.0
hyperopt==0.2.7
idna==2.10
imageio==2.9.0
imbalanced-learn==0.8.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
ipykernel==5.2.0
ipyparallel==6.2.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==5.10.1
itypes==1.2.0
jedi==0.16.0
Jinja2==2.7.3
joblib==0.14.1
jsonschema==2.4.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras==2.4.3
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.51.0
keras-tcn==3.5.0
keras-transformer==0.40.0
keras-tuner==1.3.5
kiwisolver==1.3.1
kombu==4.6.11
kt-legacy==1.0.5
lazy-object-proxy==1.7.1
Levenshtein==0.21.1
llvmlite==0.36.0
lxml==4.9.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
matplotlib-venn==0.11.6
mccabe==0.6.1
msgpack==1.0.5
munch==4.0.0
mysqlclient @ file:///tmp/build/80754af9/mysqlclient_1594308743189/work
nest-asyncio==1.5.6
networkx==2.5
numba==0.53.1
numpy==1.18.5
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==20.7
pandas==1.0.4
paramiko==2.7.1
parsl==1.0.0
parso==0.6.2
pathlib2==2.3.5
patsy==0.5.1
pegasus-wms.api==5.0.0
pegasus-wms.common==5.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
platformdirs==2.4.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.19.6
psutil==5.7.0
psycopg2==2.8.4
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
py4j==0.10.9.7
pyaml==23.5.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.1
pyflakes==2.2.0
pygame==2.4.0
pygeos==0.13
Pygments==2.6.1
pygount==1.2.4
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
PyJWT==1.7.1
pylint==2.13.4
PyMySQL==0.10.1
PyNaCl==1.3.0
pynndescent==0.5.10
pyparsing==2.4.7
pyproj==3.0.1
PySocks==1.7.1
pytest==6.1.2
pytest-celery==0.0.0a1
pytest-django==4.1.0
pytest-forked==1.3.0
pytest-xdist==2.1.0
python-dateutil==2.8.1
python-Levenshtein==0.21.1
pytz==2020.1
PyVCF==0.4.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
rapidfuzz==2.11.1
ray==2.4.0
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-allel==1.3.6
scikit-image==0.17.2
scikit-learn==0.24.2
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.4.1
seaborn==0.11.2
Shapely==1.8.5.post1
simplejson==3.17.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smote==0.1
soupsieve==2.0
sqlparse==0.4.1
statsmodels==0.12.1
tabulate==0.8.7
tblib==1.6.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.6
tensorflow==2.3.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.4.0
tensorflow-hub==0.14.0
tensorrt==8.6.1
termcolor==1.1.0
textwrap3==0.9.2
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.2
tomli==1.2.3
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typed-ast==1.5.2
typeguard==2.7.1
typing-extensions==3.7.4.3
umap-learn==0.5.3
uritemplate==3.0.1
urllib3==1.26.2
vat==0.3.3
verta==0.15.9
vine==1.3.0
virtualenv==20.17.1
wavenet==0.1.2
wcwidth==0.2.5
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
xgboost==1.5.2
yapf==0.30.0
zipp==3.4.0
